Evolution

EpiSiming.force_of_infection! — Function

Compute the force of infection on each individual, i.e. return a vector λ where each λ[i] is the force of infection on the individual with index i.

The force of infection λ[i] is computed as the sum of the force of infection in each group of direct contacts of the individual, be it their residence, their clusters (e.g. school, workplace), their network neighbors (e.g. friends, neighbors, local shops), and random contact (e.g. commuting, sporadical contacts).

For each group say i_1, ..., i_n, the associated force of contact is computed as the sum of the infectivity rate for each of the infected individuals, multiplied by the transmission rate in that group and divided by the number n-1 of co-members in the group.

EpiSiming.step_foward! — Function

step_foward!(rng, population, chances, λ, k)

A single step forward, based on the given force of infection λ (already computed with force_of_infection!).

chances is preallocated and is mutated each step with new random numbers.

The phase transitions from susceptible to exposed depends on the success of

`chances[n] ≤ 1 - exp(-λ[n] * population.susceptibility[n])`

The transition from the other states are given by [transition_rules](@ref).

EpiSiming.evolve! — Function

evolve!(
    rng, population, residences, clusters, τ, num_steps, time_step;
    verbose_step::Integer = 0
)

It essentially loops through [step_foward!](@ref) as many as num_steps times, and returns a sparse array with the transition phases ocurring for each individual.

EpiSiming.get_summary — Function

get_summary(evolution)

With m phases (SUSCEPTIBLE, EXPOSED, etc.) and num_steps iterations, get_summary(evolution) returns a num_steps x m matrix of Integers with the population count at each iteration, on each phase.