Structures

Epidemic phases for an individual

We define a type to hold the epidemic phase of an individual (e.g. asymptotic, exposed, infected, recovered, and so on). This type is defined as a primitive type which uses a single byte (8 bits), in order to reduce the memory usage. That is the smallest possible size for a primitive type. Even Bools are 8 bits. The only information entity that uses less memory is BitVector, but that is a Vector, not a singleton, and is not very performant for random access; it is more suitable for sequential access.

We could have used straight Int8 to represent the phases of the disease, but defining a new type has the advantage of allowing us to overload Base.show to display the phase in a more friendly and recognizable way. So, the SUSCEPTIBLE state can be displayed as "SUSCEPTIBLE" or "S", instead of say "0" or "1". Imagine figuring out whether "5" is RECOVERED, DECEASED or something else.

EpiSiming.Phase — Type

primitive type Phase <: Number 8 end

Phase is a primitive subtype of Number with 8 bits, representing the different phases, or stages, of a disease, for a given individual. The following phases are defined as (global) constants:

SUSCEPTIBLE = Phase(0)

EXPOSED = Phase(1)

INFECTED = Phase(2)

ASYMPTOMATIC = Phase(3)

RECOVERED = Phase(4)

DECEASED = Phase(5)

Residences

EpiSiming.Residences — Type

Residences{T, S} <: AbstractVector{Tuple{T, S}}

Residences structure with the following fields:

block::Vector{T}

position::Vector{Tuple{S, S}}

num_residents::Vector{T}

residents::Vector{Vector{T}}

Usually, T is an Int and S is a floating point type.

Population

EpiSiming.Population — Type

struct Population{R, S, T, U, V, W, X} <: AbstractVector{Tuple{{R, S, T, U, V, W, X}}}

Population structure with the following fields:

phase::Vector{R}

past_transition::Vector{S}

next_transition::Vector{Tuple{R, S}}

residence::Vector{T}

position::Vector{Tuple{U, U}}

age::Vector{V}

susceptibility::Vector{W}

infectivity::Vector{W}

clusters::Vector{Dict{X, T}}

networks::Vector{Dict{X, T}}

Usually, R=Phase, S=Int, T=U=Int, U=W=Float64, X=Symbol. If is a huge population and memory is critical, one can set S=Int16, U=Float16, V=Int8, W=Float16, X=Int8.

Clusters

EpiSiming.Clusters — Type

struct Clusters{R, S, T, U}

Clusters structure with the following fields:

id::R

name::S

contact_rate::T

clusters::Vector{Vector{U}}

Networks

This has not been implemented yet.

EpiSiming.Networks — Type

struct Networks

Structure for the set of networks

Scenario

This is not being used yet and its fields have not been defined in a suitable way.

EpiSiming.Scenario — Type

Structure for the Scenario, with the following fields:

name::String

info::String

num_population::Int

population::Population

residences::Residences

ressizedistribution::StatsBase.Weights{Float64, Float64, Vector{Float64}}

pop_pyramid::StatsBase.Weights{Float64, Float64, Vector{Float64}}

Γ_susceptibility::NamedTuple{(:shape, :scale), Tuple{Float64, Float64}}

Γ_infectivity::NamedTuple{(:shape, :scale), Tuple{Float64, Float64}}

contact_rate::NamedTuple{(:residences, :general), Tuple{Float64, Float64}}

recovery_rate::NamedTuple{ (:exposed, :infected, :asymptomatic), Tuple{Float64, Float64, Float64} }